+++ src/WebUI/Areas/Identity/Pages/Account/Manage/Address.cshtml.cs
@@ -84,8 +84,8 @@ namespace DeliveryWebApp.WebUI.Areas.Identity.Pages.Account.Manage
 
             [Required] [DataType(DataType.Text)] public string Country { get; set; }
 
-            public float Latitude { get; set; }
-            public float Longitude { get; set; }
+            public decimal Latitude { get; set; }
+            public decimal Longitude { get; set; }
         }
 
         public async Task<IActionResult> OnGetAsync()
